Break up the project package into more files
