+++ go.mod
@@ -6,6 +6,6 @@ require (
 	github.com/creack/pty v1.1.24
 	github.com/goccy/go-yaml v1.19.0
 	github.com/google/uuid v1.6.0
+	github.com/google/go-cmp v0.7.0
 )
 
-require github.com/google/go-cmp v0.7.0 // indirect
